public/requirements.txt
Django==1.7
Pillow==2.8.1
PyYAML==3.11
argh==0.25.0
astroid==1.2.1
backports.ssl-match-hostname==3.4.0.2
coverage==3.7.1
django-extensions==1.3.7
djangorestframework==3.0.4
flup==1.0.2
gnureadline==6.3.3
gunicorn==18.0
ipython==2.0.0
logilab-common==0.62.1
lxml==3.4.4
matplotlib==1.3.1
music21==1.9.3
nose==1.3.3
numpy==1.8.1
pathtools==0.1.2
psycopg2==2.5.3
pylint==1.3.1
pymei==2.0
pyparsing==2.0.2
python-dateutil==2.2
pytz==2014.4
requests==2.7.0
scipy==0.14.0
six==1.6.1
solrpy==0.9.6
tornado==3.2.2
wsgiref==0.1.2
zope.interface==4.1.1
## The following requirements were added by pip freeze:
django-markdown==0.8.4
Markdown==2.6.2
